Changed directory structure a little in pipeline components. Moved all the transform components into the new transform folder
